# Conflicts:
#	README.md

+++ vllm/v1/worker/gpu/sample/states.py
@@ -50,7 +50,9 @@ class SamplingStates:
         seed = sampling_params.seed
         self.seeds_set[req_idx] = seed is not None
         if seed is None:
-            seed = np.random.randint(_NP_INT64_MIN, _NP_INT64_MAX)
+            # Force int64: numpy's default integer dtype is int32 on Windows,
+            # which cannot represent the int64 seed range.
+            seed = np.random.randint(_NP_INT64_MIN, _NP_INT64_MAX, dtype=np.int64)
         self.seeds.np[req_idx] = seed
 
         num_logprobs = sampling_params.logprobs
